SystemUI: Clean up QS panel reordering logic
Layout and reordering has always been an important part of a display
flow. This new revolutionary change that nobody has ever seen before
will blow your mind and be amazing in all ways.

But seriously though, the change drops the old layoutTile calls in
favor of the TileAnimator's animations. The layout's also generated
a lot more based on the exact positions, not on abstract rows and
columns, most notably, due to the fact that sometimes a column count
in a single row can change which, in the old implementation, caused
jumping tiles that looked strange. Additionally, the drag events are
now handled by the QS panel directly, not through child views. This
has the benefit of making the flow easier to oversee.

Change-Id: I75f157c76cba19c591af6ada329ae402f4a12cce