SystemUI: Remove QS tile background when unused
There are times when Quick Settings tile backgrounds should be removed
since they are not expected to be there - that is when the tile is
switched to the dual state. In those instances, the top view is expected
to take over with the tile background instead.. and we don't want the
background to be set on multiple views, so it has to be removed. Simple.

Change-Id: Icebdba51f2dfc32183f68424a5fcdd797c9758d0