SystemUI: Calculate row count based on tile count
Prior to this commit, row count was based on what rows the tiles were
placed in. This works for most usual use-cases. We stumbled upon issues
however when we were trying to unhide a tile when it was the first tile
to be added (after all the tiles had been hidden). This did not show up
during testing and went undetected by us for a while.

Now we calculate the row count by looking at how many tiles we have and
splitting them across rows based on the count of tiles required to fill
up each row. This is a much more solid approach for calculating the row
count as long as we remember to call countRows after a tile's `hidden`
field changes.

Fixes issue where no tiles could not be added back to the main screen
after all the existing tiles had been hidden.

Change-Id: I2551ed26b74d6c15e8cbb177f827653162c79f15